backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pdfplumber>=0.11.4
google-genai>=1.0.0
python-dotenv>=1.0.1
